chatTest works?
it works if i run the testcases with the arrows in the indices, but the imports don't work for some reason still?

I have been at this for hours and I went to the lengths of refactoring everything into a maven project, and then i messed with the directories and got diddied.

Help me